block, bfq-sq, bfq-mq: fix asymmetric scenarios detection
Since commit e5c77a19328a ("block, bfq-sq, bfq-mq: improve asymmetric
scenarios detection"), a scenario is defined asymmetric when one of
the following conditions holds:
- active bfq_queues have different weights
- one or more group of entities (bfq_queue or other groups of entities)
  are active
bfq grants fairness and low latency also in such asymmetric scenarios,
by plugging the dispatching of I/O if the bfq_queue in service happens
to be temporarily idle. This plugging may lower throughput, so it is
important to do it only when strictly needed.

By mystake, in commit 'e5c77a19328a' ("block, bfq-sq, bfq-mq: improve
asymmetric scenarios detection") the num_active_groups counter was
firstly incremented and subsequently decremented at any entity (group
or bfq_queue) weight change.

This is useless, because only transitions from active to inactive and
vice versa matter for that counter. Unfortunately this is also
incorrect in the following case: the entity at issue is a bfq_queue
and it is under weight raising. In fact in this case there is a
spurious increment of the num_active_groups counter.

This spurious increment may cause scenarios to be wrongly detected as
asymmetric, thus causing useless plugging and loss of throughput.

This commit fixes this issue by simply removing the above useless and
wrong increments and decrements.  Logging about num_active_groups is
also improved.

Fixes: e5c77a19328a ("block, bfq-sq, bfq-mq: improve asymmetric scenarios detection")
Signed-off-by: Federico Motta <federico@willer.it>
Signed-off-by: Paolo Valente <paolo.valente@linaro.org>