leds: Force white led to luminate in offline-charging mode
Apparently, our leds aren't properly setup when the device is in offline
charging mode. We already tried userland-ways (using init.rc), but for
some weird reason it still not luminated.

So, let's hack it directly from here by filling "white" led on the dev
registration, but only when androidboot.mode is in "charger" mode.

Signed-off-by: Nauval Rizky <enuma.alrizky@gmail.com>